+++ vgg16_model.py
@@ -35,46 +35,12 @@ class vgg16(object):
         raise NotImplementedError(
             'forward_pass() is implemented in ResNet sub classes')
 
-    def inference(self, image, keep_prob):
-        # load data
-        pass
-
-        # preprocess
-        pass
-
-        # deconvolution
-        with tf.variable_scope("inference"):
-            pool3, pool4, fc8 = self._vgg16_modified(image_input, 1, 0.5, True)
-            # upsample and add with pool4
-            deconv_shape1 = pool4.get_shape()
-            W_t1 = tf.Variable(initializer=tf.truncated_normal([4, 4, deconv_shape1[3].value, 21], 0.02), name="W_t1")
-            b_t1 = tf.Variable(initializer=tf.constant(0.0, shape=[deconv_shape1[3].value]), name="b_t1")
-            conv_t1 = self.conv2d_transpose_strided(fc8, W_t1, b_t1, output_shape=tf.shape(pool4))
-            fuse_1 = tf.add(conv_t1, pool4, name="fuse_1")
-
-            # upsample and add with pool5
-            deconv_shape2 = pool3.get_shape()
-            W_t2 = tf.Variable(initializer=tf.truncated_normal([4, 4, deconv_shape2[3].value, 21], 0.02), name="W_t2")
-            b_t2 = tf.Variable(initializer=tf.constant(0.0, shape=[deconv_shape2[3].value]), name="b_t2")
-            conv_t2 = self.conv2d_transpose_strided(fuse_1, W_t2, b_t2, output_shape=tf.shape(pool3))
-            fuse_2 = tf.add(conv_t2, pool3, name="fuse_2")
-
-            shape = tf.shape(image_input)
-            deconv_shape3 = tf.stack([shape[0], shape[1], shape[2], 21])
-            W_t3 = tf.Variable(initializer=tf.truncated_normal([16, 16, 21, deconv_shape2[3].value], 0.02), name="W_t3")
-            b_t3 = tf.Variable(initializer=tf.constant(0.0, shape=[21]), name="b_t3")
-            conv_t3 = self.conv2d_transpose_strided(fuse_2, W_t3, b_t3, output_shape=deconv_shape3, stride=8)
-
-            prediction = tf.argmax(conv_t3, dimension=3, name="prediction")
-
-        return tf.expand_dims(prediction, dim=3), conv_t3
-
     def _vgg16_modified(self, x, stride, keep_prob, train=False):
         with tf.name_scope('vgg16') as name_scope:
             orig_x = x
 
-            x = self._batch_norm(x)
-            x = self._relu(x)
+            # x = self._batch_norm(x)
+            # x = self._relu(x)
 
             """ conv_1 """
             self.conv1_1 = self._conv(x, self._kernel_size, 64, self._stride, 'conv1_1')
@@ -139,8 +105,8 @@ class vgg16(object):
                 use_bias=False,
                 data_format=self._data_format)
 
-            bn = self._batch_norm(x)
-            relu = self._relu(bn)
+            # bn = self._batch_norm(x)
+            relu = self._relu(x)
             # Add summary to Tensorboard
             _activation_summary(relu)
             return relu
